Two-way CAN now part of menusystem. SRAM works, implementation for global variables or menusystem possible

+++ GccApplication1/GccApplication1/timer.h
@@ -9,4 +9,4 @@ void SysTick_Handler       ( void );
 
 uint32_t return_milliseconds();
 
-uint8_t return_seconds();
+uint16_t return_seconds();
